productivity_tips  /requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
openai>=1.0.0
